requirements.txt
fastapi
uvicorn[standard]
aiofiles==23.2.1

# Database and auth
databases[postgresql]
SQLAlchemy
asyncpg
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator

# Azure services
azure-ai-formrecognizer
azure-identity
openai

# Environment variables
python-dotenv
